07 May 2024 - Version 3.24.1.103 (14294744)

+++ Assembly-CSharp/SDG.Unturned/ClientMessageHandler_InvokeMethod.cs
@@ -17,7 +17,7 @@ internal static class ClientMessageHandler_InvokeMethod
             UnturnedLog.warn("out of bounds method index ({0}/{1})", value, NetReflection.clientMethodsLength);
             return;
         }
-        ClientMethodInfo clientMethodInfo = NetReflection.clientMethods[value];
+        ClientMethodInfo clientMethodInfo = NetReflection.clientMethods[(int)value];
         ClientInvocationContext context = new ClientInvocationContext(ClientInvocationContext.EOrigin.Remote, reader, clientMethodInfo);
         try
         {
